Split hypercube into header and source files
